requirements.txt
# Personalized Email Generator Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Streamlit for UI
streamlit==1.28.1

# OpenAI API
openai==0.28.1

# HTTP requests
requests==2.31.0

# Data validation
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# Additional utilities
python-multipart==0.0.6
